Cannot delete am insurance company when cars are assign to it

+++ MB.T.DVLD.Web/Controllers/InsuranceCompanyController.cs
@@ -16,7 +16,7 @@ namespace MB.T.DVLD.Web.Controllers
         private readonly ApplicationDbContext _context;
         private readonly IMapper _mapper;
 
-        public InsuranceCompanyController(ApplicationDbContext context , IMapper mapper)
+        public InsuranceCompanyController(ApplicationDbContext context, IMapper mapper)
         {
             _context = context;
             _mapper = mapper;
@@ -33,7 +33,7 @@ namespace MB.T.DVLD.Web.Controllers
 
             return View(insuranceCompanyVM);
         }
-      
+
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -48,7 +48,7 @@ namespace MB.T.DVLD.Web.Controllers
                 return NotFound();
             }
 
-            var insuranceCompanyVM = _mapper.Map<InsuranceCompany,InsuranceCompanyVM>(insuranceCompany);
+            var insuranceCompanyVM = _mapper.Map<InsuranceCompany, InsuranceCompanyVM>(insuranceCompany);
 
             return View(insuranceCompanyVM);
         }
@@ -56,7 +56,7 @@ namespace MB.T.DVLD.Web.Controllers
         public IActionResult Create()
         {
             return View();
-        }       
+        }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -122,15 +122,28 @@ namespace MB.T.DVLD.Web.Controllers
             }
             return View(insuranceCompany);
         }
-        
+
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var insuranceCompany = await _context.InsuranceCompanies.FindAsync(id);
-            _context.InsuranceCompanies.Remove(insuranceCompany);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            var cannotDelete = _context
+                                .Cars
+                                .Include(car => car.InsurancePolicy)
+                                .Any(car => car.InsurancePolicy.InsuranceCompanyId == id);
+
+            if (cannotDelete)
+            {
+                // TODO inform the user that he cannot delete
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                var insuranceCompany = await _context.InsuranceCompanies.FindAsync(id);
+                _context.InsuranceCompanies.Remove(insuranceCompany);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
         }
 
         #endregion
@@ -139,7 +152,7 @@ namespace MB.T.DVLD.Web.Controllers
         private bool InsuranceCompanyExists(int id)
         {
             return _context.InsuranceCompanies.Any(e => e.Id == id);
-        } 
+        }
         #endregion
     }
 }
